requirements.txt
# --- API ---
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.12

# --- Config / env ---
python-dotenv==1.0.1

# --- PDF ingestion ---
pypdf==5.1.0

# --- Text splitting ---
langchain==1.2.7
langchain-text-splitters==1.1.0

# --- Vector store ---
chromadb==0.5.23
posthog==3.5.0

# --- Gemini ---
google-generativeai==0.8.3

# --- Utils ---
numpy==2.1.3
rank-bm25==0.2.2

# --- Testing ---
pytest==8.3.4
pytest-cov==5.0.0
httpx==0.28.1
